+++ src/frameworks/Carbon/HIToolbox/CMakeLists.txt
@@ -20,6 +20,7 @@ add_framework(HIToolbox
 	SOURCES
 		src/Keyboards.cpp
 		src/Events.cpp
+		src/CarbonEventsCore.c
 		src/MacWindows.cpp
 		src/TextInputSources.mm
 		src/IBCarbonRuntime.c
