+++ src/frameworks/Carbon/src/Carbon.c
@@ -45,18 +45,6 @@ OSErr DeactivateTSMDocument(TSMDocumentID a)
 	return 0;
 }
 
-OSStatus CallNextEventHandler(EventHandlerCallRef a, EventRef b)
-{
-    if (verbose) puts("STUB: CallNextEventHandler called");
-	return 0;
-}
-
-OSStatus CreateEvent(CFAllocatorRef a, UInt32 b, UInt32 c, EventTime d, EventAttributes e, EventRef * f)
-{
-    if (verbose) puts("STUB: CreateEvent called");
-	return 0;
-}
-
 OSStatus CreateStandardAlert(AlertType a, CFStringRef b, CFStringRef c, const AlertStdCFStringAlertParamRec * d, DialogRef * e)
 {
     if (verbose) puts("STUB: CreateStandardAlert called");
@@ -80,33 +68,9 @@ EventTargetRef GetApplicationEventTarget(void)
 	return (EventTargetRef)0;
 }
 
-UInt32 GetEventClass(EventRef a)
-{
-    if (verbose) puts("STUB: GetEventClass called");
-	return 0;
-}
-
 OSStatus GetEventDispatcherTarget()
 {
     if (verbose) puts("STUB: GetEventDispatcherTarget called");
-	return 0;
-}
-
-UInt32 GetEventKind(EventRef a)
-{
-    if (verbose) puts("STUB: GetEventKind called");
-	return 0;
-}
-
-OSStatus GetEventParameter(EventRef a, EventParamName b, EventParamType c, EventParamType * d, UInt32 e, UInt32 * f, void * g)
-{
-    if (verbose) puts("STUB: GetEventParameter called");
-	return 0;
-}
-
-OSStatus GetMainEventQueue()
-{
-    if (verbose) puts("STUB: GetMainEventQueue called");
 	return 0;
 }
 
@@ -140,33 +104,15 @@ void HideMenuBar(void)
 	
 }
 
-OSStatus InstallEventHandler(EventTargetRef a, EventHandlerUPP b, UInt32 c, const EventTypeSpec * d, void * e, EventHandlerRef * f)
-{
-    if (verbose) puts("STUB: InstallEventHandler called");
-	return 0;
-}
-
 OSErr NMInstall(NMRecPtr a)
 {
     if (verbose) puts("STUB: NMInstall called");
 	return 0;
 }
 
-EventHandlerUPP NewEventHandlerUPP(EventHandlerProcPtr a)
-{
-    if (verbose) puts("STUB: NewEventHandlerUPP called");
-	return (EventHandlerUPP)0;
-}
-
 OSErr NewTSMDocument(short a, InterfaceTypeList b, TSMDocumentID * c, long d)
 {
     if (verbose) puts("STUB: NewTSMDocument called");
-	return 0;
-}
-
-OSStatus PostEventToQueue(EventQueueRef a, EventRef b, EventPriority c)
-{
-    if (verbose) puts("STUB: PostEventToQueue called");
 	return 0;
 }
 
@@ -182,24 +128,6 @@ OSStatus PutScrapFlavor(ScrapRef a, ScrapFlavorType b, ScrapFlavorFlags c, Size 
 	return 0;
 }
 
-OSStatus ReceiveNextEvent(UInt32 a, const EventTypeSpec * b, EventTimeout c, Boolean d, EventRef * e)
-{
-    if (verbose) puts("STUB: ReceiveNextEvent called");
-	return 0;
-}
-
-OSStatus ReleaseEvent(EventRef a)
-{
-    if (verbose) puts("STUB: ReleaseEvent called");
-	return 0;
-}
-
-OSStatus RemoveEventHandler(EventHandlerRef a)
-{
-    if (verbose) puts("STUB: RemoveEventHandler called");
-	return 0;
-}
-
 void RunApplicationEventLoop(void)
 {
     if (verbose) puts("STUB: RunApplicationEventLoop called");
@@ -209,12 +137,6 @@ void RunApplicationEventLoop(void)
 OSStatus RunStandardAlert(DialogRef a, ModalFilterUPP b, DialogItemIndex * c)
 {
     if (verbose) puts("STUB: RunStandardAlert called");
-	return 0;
-}
-
-OSStatus SendEventToEventTarget(EventRef a, EventTargetRef b)
-{
-    if (verbose) puts("STUB: SendEventToEventTarget called");
 	return 0;
 }
